requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
dlib>=19.24.0
matplotlib>=3.7.0 